requirements.txt
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
gymnasium>=0.28.1
pybullet>=3.2.5
stable-baselines3>=2.2.1
tqdm>=4.66.0
tensorboard>=2.14.0
opencv-python>=4.8.0
scipy>=1.11.0
seaborn>=0.12.2
